Refactor: pull out LayoutFooter
Finally, let's pull out the LayoutFooter and just export that in
`index`.

In the next commit(s) we will remove unused imports and clean up
the files.
